merging

+++ controllers/InternshipController.php
@@ -2,7 +2,6 @@
 
 namespace app\controllers;
 
-
 use Yii;
 use app\models\Internship;
 use app\models\SearchInternship;
@@ -37,22 +36,21 @@ class InternshipController extends Controller
      */
     public function actionIndex()
     {
-    
-    if(!Yii::$app->user->isGuest){
-        $searchModel = new SearchInternship();
-        if(Yii::$app->request->get('from') && Yii::$app->request->get('to')){
-            $searchModel->to = Yii::$app->request->get('to');
-            $searchModel->from = Yii::$app->request->get('from');
-        }
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+        if (!Yii::$app->user->isGuest) {
+            $searchModel = new SearchInternship();
+            if (Yii::$app->request->get('from') && Yii::$app->request->get('to')) {
+                $searchModel->to = Yii::$app->request->get('to');
+                $searchModel->from = Yii::$app->request->get('from');
+            }
+            $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        return $this->render('index', [
+            return $this->render('index', [
             'searchModel' => $searchModel,
             'dataProvider' => $dataProvider,
         ]);
-    }else{
-        throw new \yii\web\ForbiddenHttpException;
-    }
+        } else {
+            throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
@@ -63,13 +61,13 @@ class InternshipController extends Controller
      */
     public function actionView($id)
     {
-        if(!Yii::$app->user->isGuest){
+        if (!Yii::$app->user->isGuest) {
             return $this->render('view', [
                 'model' => $this->findModel($id),
             ]);
-        }else{
+        } else {
             throw new \yii\web\ForbiddenHttpException;
-        }    
+        }
     }
 
     /**
@@ -81,37 +79,66 @@ class InternshipController extends Controller
     {
         $model = new Internship();
 
-        if(!Yii::$app->user->isGuest){
-            if ($model->load(Yii::$app->request->post()) ){
-            $model->file = UploadedFile::getInstance($model, 'file');
-            $model->file1 = UploadedFile::getInstance($model, 'file1');
-            $model->file2 = UploadedFile::getInstance($model, 'file2');
-            $model->file3 = UploadedFile::getInstance($model, 'file3');
-                if ($model->file ) {                
-                    $model->file->saveAs('uploads/internship/' . $model->file ->baseName . '.' . $model->file ->extension);
-                    $model->file= 'uploads/internship/' . $model->file ->baseName . '.' . $model->file ->extension;
-                }
-                if ($model->file1 ) {                
-                    $model->file1->saveAs('uploads/internship/' . $model->file1 ->baseName . '.' . $model->file1 ->extension);
-                    $model->file1= 'uploads/internship/' . $model->file1 ->baseName . '.' . $model->file1 ->extension;
-                }
-                if ($model->file2 ) {                
-                    $model->file2->saveAs('uploads/internship/' . $model->file2 ->baseName . '.' . $model->file2 ->extension);
-                    $model->file2= 'uploads/internship/' . $model->file2 ->baseName . '.' . $model->file2 ->extension;
-                }
-                if ($model->file3 ) {                
-                    $model->file3->saveAs('uploads/internship/' . $model->file3 ->baseName . '.' . $model->file3 ->extension);
-                    $model->file3= 'uploads/internship/' . $model->file3 ->baseName . '.' . $model->file3 ->extension;
-                }
+        if (!Yii::$app->user->isGuest) {
+            if ($model->load(Yii::$app->request->post())) {
+                $model->file = UploadedFile::getInstance($model, 'file');
+                $model->file1 = UploadedFile::getInstance($model, 'file1');
+                $model->file2 = UploadedFile::getInstance($model, 'file2');
+                $model->file3 = UploadedFile::getInstance($model, 'file3');
                 
-	            $model->save(false);
+                if ($model->file) {
+                    $cnt = 1;
+                    $filename =  'uploads/internship/' . $model->file ->baseName . '.' . $model->file ->extension;
+                    while (file_exists($filename)) {
+                        $filename =  'uploads/internship/' . $model->file ->baseName. $cnt . '.' . $model->file ->extension ;
+                        $cnt++;
+                    }
+                    $model->file->saveAs($filename);
+                    $model->file= $filename;
+                }
+                if ($model->file1) {
+                    $cnt = 1;
+                    $filename =  'uploads/internship/' . $model->file1 ->baseName . '.' . $model->file1 ->extension;
+                    while (file_exists($filename)) {
+                        $filename =  'uploads/internship/' . $model->file1 ->baseName. $cnt . '.' . $model->file1 ->extension ;
+                        $cnt++;
+                    }
+                    $model->file1->saveAs($filename);
+                    $model->file1= $filename;
+                }
+                if ($model->file2) {
+                    $cnt = 1;
+                    $filename =  'uploads/internship/' . $model->file2 ->baseName . '.' . $model->file2 ->extension;
+                    while (file_exists($filename)) {
+                        $filename =  'uploads/internship/' . $model->file2 ->baseName. $cnt . '.' . $model->file2 ->extension ;
+                        $cnt++;
+                    }
+                    $model->file2->saveAs($filename);
+                    $model->file2= $filename;
+                }
+                if ($model->file3) {
+                    $cnt = 1;
+                    $filename =  'uploads/internship/' . $model->file3 ->baseName . '.' . $model->file3 ->extension;
+                    while (file_exists($filename)) {
+                        $filename =  'uploads/internship/' . $model->file3 ->baseName. $cnt . '.' . $model->file3 ->extension ;
+                        $cnt++;
+                    }
+                    $model->file3->saveAs($filename);
+                    $model->file3= $filename;
+                }
+                $model->user_id = Yii::$app->user->id;
+                $model->save(false);
+                //  echo $model->seminar_id;
                 return $this->redirect(['view', 'id' => $model->internship_id]);
             }
 
-        return $this->render('create', [
+            return $this->render('create', [
             'model' => $model,
         ]);
-    }}
+        } else {
+            throw new \yii\web\ForbiddenHttpException;
+        }
+    }
 
     /**
      * Updates an existing Internship model.
@@ -125,52 +152,75 @@ class InternshipController extends Controller
         $model = $this->findModel($id);
         $old_data= $this->findModel($id);
         
-        if(!Yii::$app->user->isGuest){
-        if ($model->load(Yii::$app->request->post())) {
-            $model->file = UploadedFile::getInstance($model, 'file');
+        if (!Yii::$app->user->isGuest) {
+            $model = $this->findModel($id);
+            $old_data= $this->findModel($id);
+    
+            if ($model->load(Yii::$app->request->post())) {
+                $model->file = UploadedFile::getInstance($model, 'file');
+                
+                if (!$model->file) {
+                    $model->file = $old_data->file;
+                } else {
+                    $cnt = 1;
+                    $filename =  'uploads/internship/' . $model->file ->baseName . '.' . $model->file ->extension;
+                    while (file_exists($filename)) {
+                        $filename =  'uploads/internship/' . $model->file ->baseName. $cnt . '.' . $model->file ->extension ;
+                        $cnt++;
+                    }
+                    $model->file->saveAs($filename);
+                    $model->file= $filename;
+                }
+                $model->file1 = UploadedFile::getInstance($model, 'file1');
+                if (!$model->file1) {
+                    $model->file1 = $old_data->file1;
+                } else {
+                    $cnt = 1;
+                    $filename =  'uploads/internship/' . $model->file1 ->baseName . '.' . $model->file1 ->extension;
+                    while (file_exists($filename)) {
+                        $filename =  'uploads/internship/' . $model->file1 ->baseName. $cnt . '.' . $model->file1 ->extension ;
+                        $cnt++;
+                    }
+                    $model->file1->saveAs($filename);
+                    $model->file1= $filename;
+                }
+                $model->file2 = UploadedFile::getInstance($model, 'file2');
+                if (!$model->file2) {
+                    $model->file2 = $old_data->file2;
+                } else {
+                    $cnt = 1;
+                    $filename =  'uploads/internship/' . $model->file2 ->baseName . '.' . $model->file2 ->extension;
+                    while (file_exists($filename)) {
+                        $filename =  'uploads/internship/' . $model->file2 ->baseName. $cnt . '.' . $model->file2 ->extension ;
+                        $cnt++;
+                    }
+                    $model->file2->saveAs($filename);
+                    $model->file2= $filename;
+                }
+                $model->file3 = UploadedFile::getInstance($model, 'file3');
+                
+                if (!$model->file3) {
+                    $model->file3 = $old_data->file3;
+                } else {
+                    $cnt = 1;
+                    $filename =  'uploads/internship/' . $model->file3 ->baseName . '.' . $model->file3 ->extension;
+                    while (file_exists($filename)) {
+                        $filename =  'uploads/internship/' . $model->file3 ->baseName. $cnt . '.' . $model->file3 ->extension ;
+                        $cnt++;
+                    }
+                    $model->file3->saveAs($filename);
+                    $model->file3= $filename;
+                }
+                $model->save(false);
             
-            if (!$model->file){
-                $model->file = $old_data->file;
-            }else{
-                
-                $model->file->saveAs('uploads/internship/' . $model->file ->baseName . '.' . $model->file ->extension);
-                $model->file= 'uploads/internship/' . $model->file ->baseName . '.' . $model->file ->extension;
+                return $this->redirect(['view', 'id' => $model->internship_id]);
             }
-            $model->file1 = UploadedFile::getInstance($model, 'file1');
-            
-            if (!$model->file1 ){
-                $model->file1 = $old_data->file1;
-                
-            }else{
-                
-                $model->file1->saveAs('uploads/internship/' . $model->file1 ->baseName . '.' . $model->file1 ->extension);
-                $model->file1= 'uploads/internship/' . $model->file1 ->baseName . '.' . $model->file1 ->extension;
-            }
-            $model->file2 = UploadedFile::getInstance($model, 'file2');
-            if (!$model->file2){
-                $model->file2 = $old_data->file2;
 
-            }else{
-                
-                $model->file2->saveAs('uploads/internship/' . $model->file2 ->baseName . '.' . $model->file2 ->extension);
-                $model->file2= 'uploads/internship/' . $model->file2 ->baseName . '.' . $model->file2 ->extension;
-            }
-            $model->file3 = UploadedFile::getInstance($model, 'file3');
-            if (!$model->file3){
-                $model->file3 = $old_data->file3;
-
-            }else{
-                $model->file3->saveAs('uploads/internship/' . $model->file3 ->baseName . '.' . $model->file3 ->extension);
-                $model->file3= 'uploads/internship/' . $model->file3 ->baseName . '.' . $model->file3 ->extension;
-            }
-            $model->save(false);
-            
-            return $this->redirect(['view', 'id' => $model->internship_id]);
-        }
-
-        return $this->render('update', [
+            return $this->render('update', [
             'model' => $model,
         ]);
+        } else {
+            throw new \yii\web\ForbiddenHttpException;
         }
     }
 
@@ -184,9 +234,18 @@ class InternshipController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        // $model = $this->findModel($id);
+        // unlink($model->file);
+        // unlink($model->file1);
+        // unlink($model->file2);
+        // unlink($model->file3);
+        if (!Yii::$app->user->isGuest) {
+            $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+            return $this->redirect(['index']);
+        } else {
+            throw new \yii\web\ForbiddenHttpException;
+        }
     }
 
     /**
